Introduce a SQLite cache to amortize the cost of computing docs for 3rd party crates
